+++ packages/xprog-prensa/primitives/Accordion/component.tsx
@@ -10,6 +10,9 @@ export const Accordion: React.FC<AccordionProps> = ({
   css,
   items,
   innerSpace,
+  icon,
+  iconExpandedState,
+  iconNotExpandedState,
   size,
 }) => {
   const [state, setState] = useState(items || [])
@@ -52,6 +55,9 @@ export const Accordion: React.FC<AccordionProps> = ({
           key={id}
           id={id}
           innerSpace={innerSpace}
+          icon={otherProps.icon ? otherProps.icon : icon}
+          iconExpandedState={otherProps.iconExpandedState ? otherProps.iconExpandedState : iconExpandedState}
+          iconNotExpandedState={otherProps.iconNotExpandedState ? otherProps.iconNotExpandedState : iconNotExpandedState}
           onClick={onClick ? onClick : () => handleClick(id)}
           {...otherProps}
         />